requirements.txt
beautifulsoup4==4.12.2
img2pdf==0.5.0
Pillow==10.1.0
pycryptodome==3.19.0
Requests==2.31.0
